--- @connect-shared/lib/projects/form.ts
@@ -1,29 +1,0 @@
-import * as yup from 'yup';
-
-export const CATEGORIES = ['CeFi', 'Cross Chain', 'DeFi', 'Governance', 'NFT', 'Social', 'Utility'] as const;
-
-export type ProjectCategory = (typeof CATEGORIES)[number];
-
-export const schema = yup.object({
-  id: yup.string(),
-  name: yup.string().required('Project name is required'),
-  description: yup.string().required('Project description is required'),
-  avatar: yup.string(),
-  coverImage: yup.string(),
-  category: yup.string().oneOf(CATEGORIES).required(),
-  websites: yup.array(yup.string().url()).min(1),
-  farcasterValues: yup.array(yup.string()),
-  github: yup.string(),
-  twitter: yup.string(),
-  mirror: yup.string(),
-  projectMembers: yup
-    .array(
-      yup.object({
-        farcasterId: yup.number().required()
-      })
-    )
-    .required()
-    .min(1)
-});
-
-export type FormValues = yup.InferType<typeof schema>;
