+++ apps/connect/lib/projects/createProjectAction.ts
@@ -3,9 +3,9 @@
 import { log } from '@charmverse/core/log';
 import { authActionClient } from '@connect-shared/lib/actions/actionClient';
 import { storeProjectMetadataAndPublishOptimismAttestation } from '@connect-shared/lib/attestations/storeProjectMetadataAndPublishOptimismAttestation';
-import { createOptimismProject } from '@connect-shared/lib/projects/createOptimismProject';
-import { schema } from '@connect-shared/lib/projects/form';
+import { createProject } from '@connect-shared/lib/projects/createProject';
 import { generateOgImage } from '@connect-shared/lib/projects/generateOgImage';
+import { schema } from '@connect-shared/lib/projects/projectSchema';
 import { isTestEnv } from '@root/config/constants';
 import { charmverseProjectDataChainId, disableCredentialAutopublish } from '@root/lib/credentials/constants';
 import { storeCharmverseProjectMetadata } from '@root/lib/credentials/reputation/storeCharmverseProjectMetadata';
@@ -16,7 +16,7 @@ export const actionCreateProject = authActionClient
   .action(async ({ parsedInput, ctx }) => {
     const input = parsedInput;
     const currentUserId = ctx.session.user!.id;
-    const newProject = await createOptimismProject({
+    const newProject = await createProject({
       userId: currentUserId,
       input,
       source: 'connect'
