+++ apps/sunnyawards/lib/projects/editProjectAction.ts
@@ -3,14 +3,14 @@
 import { log } from '@charmverse/core/log';
 import { authActionClient } from '@connect-shared/lib/actions/actionClient';
 import { storeUpdatedProjectMetadataAttestation } from '@connect-shared/lib/attestations/storeUpdatedProjectMetadataAttestation';
-import type { EditOptimismProjectValues } from '@connect-shared/lib/projects/editOptimismProject';
-import { editOptimismProject } from '@connect-shared/lib/projects/editOptimismProject';
+import type { EditProjectValues } from '@connect-shared/lib/projects/editProject';
+import { editProject } from '@connect-shared/lib/projects/editProject';
 import { generateOgImage } from '@connect-shared/lib/projects/generateOgImage';
 import { charmverseProjectDataChainId, disableCredentialAutopublish } from '@root/lib/credentials/constants';
 import { storeCharmverseProjectMetadata } from '@root/lib/credentials/reputation/storeCharmverseProjectMetadata';
 import { revalidatePath } from 'next/cache';
 
-import { schema } from './form';
+import { schema } from './schema';
 
 export const editProjectAction = authActionClient
   .metadata({ actionName: 'create-project' })
@@ -18,9 +18,9 @@ export const editProjectAction = authActionClient
   .action(async ({ parsedInput, ctx }) => {
     const input = parsedInput;
     const currentUserId = ctx.session.user!.id;
-    const editedProject = await editOptimismProject({
+    const editedProject = await editProject({
       userId: currentUserId,
-      input: { ...input, projectId: input.id } as EditOptimismProjectValues
+      input: { ...input, projectId: input.id } as EditProjectValues
     });
 
     await generateOgImage(editedProject.id, currentUserId);
