+++ lib/credentials/__tests__/mapProjectToGitcoin.spec.ts
@@ -11,7 +11,8 @@ describe('mapProjectToGitcoin', () => {
       id: '1',
       name: 'Project X',
       createdBy,
-      category: 'DeFi',
+      optimismCategory: null,
+      sunnyAwardsCategory: null,
       description: 'A sample project',
       avatar: 'avatar.png',
       coverImage: 'cover.jpg',
